flyttet saveimage logikken fra kontrolleren -> services for å holde business logic unna
